large-media-converter/lambdas/requirements.txt
pydantic<2
sentry-sdk
urllib3<2
tenacity
Pillow
cryptography<39